% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Khat_RSS.R
\name{EMPL_RSS}
\alias{EMPL_RSS}
\alias{Khat_RSS}
\title{Estimation of Mean Past Lifetime under RSS}
\usage{
Khat_RSS(t, RSSdata, k, m)
}
\arguments{
\item{t}{A numeric vector}

\item{RSSdata}{Data generated from Ranked Set Sampling Scheme}

\item{k}{Collection size}

\item{m}{Number of cycles}
}
\value{
The Estimation of MPL based on the RSS
}
\description{
Estimation of Mean Past Lifetime Based on the Ranked Set Sampling Scheme (EMPL_RSS)
}
\examples{
k <- 5
m <- 5
lambda <- 1
RSSdata <- matrix(0, nrow = k, ncol = m)
for (j in 1:m) {
  for (i in 1:k) {
    x <- rgamma(k, shape = 2, scale = 1)
    y <- lambda * ((x - mean(x)) / sd(x)) + sqrt((1 - lambda^2)) * rnorm(k, 0, 1)
    sy <- sort(y)
    syindex <- sort(y, index.return = TRUE)$ix
    RSSdata[i, j] <- x[syindex][i]
  }
}
Khat_RSS(c(1, 2, 4, 5), RSSdata, k = 5, m = 5)
}
