% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mhat_SRS.R
\name{Mhat_SRS}
\alias{Mhat_SRS}
\title{Estimation of Mean Residual Lifetime under SRS}
\usage{
Mhat_SRS(t, data)
}
\arguments{
\item{t}{A numeric vector}

\item{data}{Data generated from Simple random sampling}
}
\value{
The Estimation of MRL based on the SRS
}
\description{
Estimation of Mean Residual Lifetime Based on the Simple Random Sampling Scheme (EMRL_SRS)
}
\examples{
data <- rgamma(10000, shape = 2, scale = 1)
t <- seq(0, 10, 0.1)
Mhat_SRS(t, data)
plot(t, Mhat_SRS(t, data), type = "l", col = 2, ylim = c(0, 3))
}
